refactor: move utils file into test
make sure it's ignored using jest instead of typescript
