+++ packages/react-instantsearch-hooks-web/src/widgets/__tests__/all-widgets.test.tsx
@@ -2,7 +2,7 @@
  * @jest-environment node
  */
 
-import { getAllInstantSearchWidgets } from '../__testutils__/all-widgets';
+import { getAllInstantSearchWidgets } from './__utils__/all-widgets';
 
 describe('widgets', () => {
   const widgets = getAllInstantSearchWidgets();
