+++ tsconfig.declaration.json
@@ -5,5 +5,5 @@
     "emitDeclarationOnly": true,
     "noEmit": false
   },
-  "exclude": ["**/__tests__/**/*", "**/__testutils__/**/*", "**/dist/**/*"],
+  "exclude": ["**/__tests__/**/*", "**/dist/**/*"],
 }
